Removed further scenario bar functions into the relevant file and fixed an issue with screen size adjustment file.
